Fixes according to the comments
