+++ src/components/types/types.js
@@ -37,6 +37,8 @@ export const commentType = PropTypes.shape({
   date: PropTypes.instanceOf(Date).isRequired
 }).isRequired;
 
+const commentsType = PropTypes.objectOf(commentType).isRequired;
+
 const movieType = {
   movie: movieObjectType
 };
@@ -96,14 +98,17 @@ export const ScreenTabType = {
 
 
 export const smallMovieCardType = Object.assign({}, onMouseEnterType, onMouseLeaveType, movieObjectType);
-export const promotedMovieAndMoviesType = Object.assign({}, movieObjectType, moviesType);
+export const promotedMovieMoviesCommentsType = Object.assign({}, movieObjectType, moviesType, commentsType);
 
 export const mainPageType = Object.assign({}, onUserIconClickType, onPlayButtonClickType, promotedMovieType, moviesType);
+export const moviePageType = Object.assign({}, onUserIconClickType, onPlayButtonClickType, promotedMovieType, moviesType, commentsType);
 export const movieCardTopType = Object.assign({}, onUserIconClickType, onPlayButtonClickType, movieType);
 export const headerType = Object.assign({}, onUserIconClickType, headerSettingsType);
 export const userBlockType = Object.assign({}, onUserIconClickType, isRenderedType);
 export const catalogType = Object.assign({}, headingType, moviesType, optionalGenresType);
 export const moreButtonType = Object.assign({}, headingType, moviesType);
 export const myListPageType = Object.assign({}, onUserIconClickType, moviesType);
-export const movieCardType = Object.assign({}, onUserIconClickType, onPlayButtonClickType, promotedMovieType);
-export const tabsType = Object.assign({}, ScreenTabType, movieObjectType);
+export const movieCardType = Object.assign({}, onUserIconClickType, onPlayButtonClickType, promotedMovieType, commentsType);
+export const tabsType = Object.assign({}, ScreenTabType, movieObjectType, commentsType);
+export const tabType = Object.assign({}, movieObjectType);
+export const tabWithCommentsType = Object.assign({}, movieObjectType, commentsType);
