+++ src/utils/utils.js
@@ -1,4 +1,5 @@
-import {GENRES, MOVIES, VIDEO_URLS, DIRECTORS, ACTORS, MOVIES_COUNT, MAXIMUM_RATING, MAXIMUM_RATING_COUNT, FISH_TEXT, ActorsCount, Description, Paragraph, Duration, Id, CommentCount, MAXIMUM_SIMIALAR_MOVIES} from "./const";
+import React from "react";
+import {GENRES, MOVIES, VIDEO_URLS, DIRECTORS, ACTORS, MOVIES_COUNT, MAXIMUM_RATING, MAXIMUM_RATING_COUNT, FISH_TEXT, ActorsCount, Description, Paragraph, Duration, Id, CommentCount, MAXIMUM_SIMIALAR_MOVIES, MAXIMUM_DISPLAYED_ACTORS, RatingSystem} from "./const";
 
 const getRandomElementFromArray = (array) => {
   const max = array.length;
@@ -96,12 +97,11 @@ const generateComments = (count) => {
 };
 
 export const generateCommentsById = (movieIds) => {
-  let comments = {};
-
-  movieIds.map((id) => {
+  const comments = movieIds.reduce((accumulator, id) => {
     const number = getRandomIntegerNumber(CommentCount.MINIMUM, CommentCount.MAXIMUM);
-    comments[id] = generateComments(number);
-  });
+    accumulator[id] = generateComments(number);
+    return accumulator;
+  }, {});
 
   return comments;
 };
@@ -133,4 +133,58 @@ export const filterMoviesByGenre = (movies, genre) => {
   }
 
   return finalArray;
+};
+
+export const generateDurationString = (duration) => {
+  return `${Math.floor(duration / 60)}h ${duration % 60}m`;
+};
+
+export const generateActorsText = (actors) => {
+  const actorsMaxIndex = actors.length - 1;
+
+  return (
+    <span className="movie-card__details-value">
+      {actors.map((actor, i) => (
+        <React.Fragment key={i}>
+          {actor}{i < actorsMaxIndex ? `, ` : ``} {i < actorsMaxIndex ? <br/> : ``}
+        </React.Fragment>
+      ))}
+    </span>);
+};
+
+export const generateStarringString = (actors) => {
+  let text = ``;
+  let end = ``;
+  let actorsLocal = actors.slice();
+
+  if (actorsLocal.length > MAXIMUM_DISPLAYED_ACTORS) {
+    actorsLocal = actors.slice(0, MAXIMUM_DISPLAYED_ACTORS);
+
+    end = ` and other`;
+  }
+
+  text = actorsLocal.join(`, `);
+
+  text = text.slice(0, -2);
+
+  text += end;
+
+  return text;
+};
+
+export const generateRatingText = (rating) => {
+  switch (true) {
+    case rating <= RatingSystem.BAD:
+      return `Bad`;
+    case rating <= RatingSystem.NORMAL:
+      return `Normal`;
+    case rating <= RatingSystem.GOOD:
+      return `Good`;
+    case rating < RatingSystem.AWESOME:
+      return `Very good`;
+    case rating === RatingSystem.AWESOME:
+      return `Awesome`;
+    default:
+      return `Rating is incorrect`;
+  }
 };
